requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
streamlit
tensorflow
keras
python-dotenv